+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - [iOS] Bump min supported iOS 16
 - [iOS] Migrate from Cocoapods to SPM
 - [iOS] Update Google Maps to 10.13.0
+- [Android] Migrate alerts to `Media3` library